+++ src/md_io/DecompWriter.cpp
@@ -37,7 +37,7 @@ void DecompWriter::doOutput(ParticleContainer* particleContainer, DomainDecompBa
     if(_incremental) {
       unsigned long temp = simstep/_writeFrequency;
       filenamestream << "-";
-      while(temp < floor(_numberOfTimesteps/_writeFrequency)){
+      while(temp < floor((double) (_numberOfTimesteps/_writeFrequency))){
         filenamestream << "0";
         temp = temp*10;
       }
